% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICsims-package.R
\docType{package}
\name{ICsims-package}
\alias{ICsims}
\alias{ICsims-package}
\title{Simulation framework allowing for between-sample heterogeneity in GLMs}
\description{
Simulation framework allowing for between-sample heterogeneity in GLMs
}
\details{
ICsims implements a simulation framework sampling repeatedly from linear
regression models and GLMs, allowing for between-sample heterogeneity.
The purpose is to allow the study of AIC and related statistics in the
context of model selection, with prediction quality as target.

\tabular{ll}{
Package: \tab ICsims\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2015-10-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

\code{\link{simulation.study}} is the main workhorse for simulation
}
\author{
Mark J Brewer \email{Mark.Brewer@bioss.ac.uk}
}
\references{
\url{https://github.com/MarkJBrewer/ICsims}
}
\keyword{package}

