% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noisecor.R
\name{noisecor}
\alias{noisecor}
\title{Add noise to any user specified correlation matrix}
\usage{
noisecor(cormat, epsilon = 0.01, eidim = 2)
}
\arguments{
\item{cormat}{A correlation matrix. The function does not itself check whether this is a valid
correlation matrix.}

\item{epsilon}{Numeric. The maximim perturbation of any correlation.}

\item{eidim}{Integer. The shape of the perturbation; if \code{eidim==1}, the perturbation is exactly
epsilon but in random directions. Increasing values of eidim tend to centre the perturbations.}
}
\description{
\code{noisecor} adds noise to any user specified correlation matrix, in the manner of Hardin et al (2013).
}
\references{
For full details, see
\cite{Hardin, J., Garcia, S. R., and Golan, D. (2013). A method for generating realistic correlation matrices. Annals of Applied Statistics, 7(3):1733-1762.}
This function is not intended for use by the user.
}

